# modules/mod-002-api-design/exercise-01/openapi.yaml
openapi: 3.1.0
info: { title: ML Training Jobs API, version: "1.0.0" }
servers: [{ url: https://api.ml-platform.example.com/v1 }]
paths:
  /training-jobs:
    post:
      operationId: submitJob
      parameters:
        - { name: X-Tenant-Id, in: header, required: true, schema: { type: string } }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/JobRequest' } } }
      responses:
        '202': { content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
        '429': { description: Rate limit exceeded }
    get:
      operationId: listJobs
      parameters:
        - { name: X-Tenant-Id, in: header, required: true, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
  /training-jobs/{id}:
    get:
      operationId: getJob
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    delete:
      operationId: cancelJob
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
  /training-jobs/{id}:retry:
    post:
      operationId: retryJob
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
components:
  schemas:
    JobRequest:
      type: object
      required: [model_uri, dataset_uri]
      properties:
        model_uri: { type: string, format: uri }
        dataset_uri: { type: string, format: uri }
        gpu_count: { type: integer, minimum: 1, default: 1 }
        max_runtime_hours: { type: integer, minimum: 1, maximum: 168, default: 8 }
        hyperparams: { type: object, additionalProperties: true }
    Job:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [pending, scheduled, running, succeeded, failed, cancelled] }
        spec: { $ref: '#/components/schemas/JobRequest' }
        created_at: { type: string, format: date-time }
        run_id: { type: string }
        mlflow_run_url: { type: string, format: uri }